test: Add deleteCiliumDS
[ upstream commit b5f6898ff533fea4f8b98e8c97b7f5fc6a801884 ]

They will be used by NodePort BPF tests, so move to a common module.
